+++ sqlparser/lib/src/engine/options.dart
@@ -29,13 +29,13 @@ class EngineOptions {
   EngineOptions({
     this.useMoorExtensions = false,
     this.enabledExtensions = const [],
-    this.version = SqliteVersion.current,
+    this.version = SqliteVersion.minimum,
   }) {
-    if (version.compareTo(SqliteVersion.minimum) < 0) {
+    if (version < SqliteVersion.minimum) {
       throw ArgumentError.value(
           version, 'version', 'Must at least be ${SqliteVersion.minimum}');
     }
-    if (version.compareTo(SqliteVersion.current) > 0) {
+    if (version > SqliteVersion.current) {
       throw ArgumentError.value(
           version, 'version', 'Must at most be ${SqliteVersion.current}');
     }
@@ -60,6 +60,21 @@ class EngineOptions {
 /// than the desired version.
 @sealed
 class SqliteVersion implements Comparable<SqliteVersion> {
+  /// The minimum sqlite version assumed by the current version of the
+  /// `sqlparser` package.
+  ///
+  /// This does not mean that older version aren't supported, but this library
+  /// can't provide analysis warnings when using recent sqlite3 features.
+  static const SqliteVersion minimum = SqliteVersion.v3(34);
+
+  /// Version `3.35.0` of `sqlite3`.
+  static const SqliteVersion v3_35 = SqliteVersion.v3(35);
+
+  /// The highest sqlite version supported by this `sqlparser` package.
+  ///
+  /// Newer features in `sqlite3` may not be recognized by this library.
+  static const SqliteVersion current = v3_35;
+
   /// The major version of sqlite.
   ///
   /// This will always be `3` in the foreseeable future.
@@ -79,17 +94,8 @@ class SqliteVersion implements Comparable<SqliteVersion> {
 
   const SqliteVersion.v3(int minor, [int patch = 0]) : this(3, minor, patch);
 
-  /// The minimum sqlite version assumed by the current version of the
-  /// `sqlparser` package.
-  ///
-  /// This does not mean that older version aren't supported, but this library
-  /// can't provide analysis warnings when using recent sqlite3 features.
-  static const SqliteVersion minimum = SqliteVersion.v3(34);
-
-  /// The highest sqlite version supported by this `sqlparser` package.
-  ///
-  /// Newer features in `sqlite3` may not be recognized by this library.
-  static const SqliteVersion current = SqliteVersion.v3(34);
+  bool operator <(SqliteVersion other) => compareTo(other) < 0;
+  bool operator >(SqliteVersion other) => compareTo(other) > 0;
 
   @override
   int compareTo(SqliteVersion other) {