+++ sqlparser/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.14.1-dev
 
 - Support `FROM` clauses in `UPDATE` statements
+- Support `MATERIALIZED`/`NOT MATERIALIZED` hints in common table expressions
 - Add `EngineOptions.version` argument to specify the desired sqlite version.
   Using newer features will be reported as analysis warnings.
 - Fix `rank` columns of fts5 tables being misreported as integers